# Conflicts:
#	src/main/java/com/wisein/wiselab/common/paging/AbstractPagingCustom.java
#	src/main/java/com/wisein/wiselab/common/paging/PaginationInfo.java
#	src/main/java/com/wisein/wiselab/controller/qaController.java
#	src/main/java/com/wisein/wiselab/controller/testController2.java
#	src/main/java/com/wisein/wiselab/dao/QaListDAO.java
#	src/main/java/com/wisein/wiselab/dto/QaListDTO.java
#	src/main/java/com/wisein/wiselab/service/QaListServiceImpl.java
#	src/main/resources/mapper/qaListMapper.xml
#	src/main/webapp/WEB-INF/tiles/common/header.jsp
#	src/main/webapp/WEB-INF/views/cmn/foodDetail.jsp
#	src/main/webapp/WEB-INF/views/cmn/qaList.jsp
#	src/main/webapp/resources/css/common.css
#	src/main/webapp/resources/css/foodDetail.css

+++ src/main/java/com/wisein/wiselab/dto/RecmDTO.java
@@ -9,7 +9,7 @@ public class RecmDTO {
     private int num;
     private int refMatzip;
     private String writer;
-    private int star;
+    private float star;
     private String subject;
     private String content;
     private Date regDate;
